Jaro-Winkler bug fix: A compared to B does not equal B compared to A
Currently, string A compared to string B will yield a different result to string B compared to string A, when either string A or B contains a character found in both strings, that is found more than once in one of the string. This is incorrect. The Jaro-Winkler method should yield the same result when comparing the match value from A compared to B with B compared to A.

To rectify this, when identifying the common characters, the same character should not be repeated. The common characters variable needs to be deduplicated before returned.

This fix replaces the common characters string with an array that uses the common character as the key, to avoid duplication. By running the updated method, A compared to B yields the same results as B compared to A.

This is inline with the C# version of the method.